compute: move logic to enterprise directory
