+++ cmd/frontend/internal/cli/serve_cmd.go
@@ -263,6 +263,7 @@ func Main(enterpriseSetupHook func(db database.DB, c conftypes.UnifiedWatchable)
 		enterprise.SearchContextsResolver,
 		enterprise.OrgRepositoryResolver,
 		enterprise.NotebooksResolver,
+		enterprise.ComputeResolver,
 	)
 	if err != nil {
 		return err
