+++ enterprise/cmd/frontend/internal/batches/resolvers/resolver_test.go
@@ -1903,5 +1903,5 @@ mutation($batchChange: ID!, $changesets: [ID!]!, $draft: Boolean!) {
 func stringPtr(s string) *string { return &s }
 
 func newSchema(db database.DB, r graphqlbackend.BatchChangesResolver) (*graphql.Schema, error) {
-	return graphqlbackend.NewSchema(db, r, nil, nil, nil, nil, nil, nil, nil, nil, nil)
+	return graphqlbackend.NewSchema(db, r, nil, nil, nil, nil, nil, nil, nil, nil, nil, nil)
 }
