+++ enterprise/cmd/frontend/internal/codemonitors/resolvers/resolvers_test.go
@@ -326,7 +326,7 @@ func TestQueryMonitor(t *testing.T) {
 	_, err = r.insertTestMonitorWithOpts(ctx, t, actionOpt, postHookOpt)
 	require.NoError(t, err)
 
-	schema, err := graphqlbackend.NewSchema(db, nil, nil, nil, nil, r, nil, nil, nil, nil, nil)
+	schema, err := graphqlbackend.NewSchema(db, nil, nil, nil, nil, r, nil, nil, nil, nil, nil, nil)
 	require.NoError(t, err)
 
 	t.Run("query by user", func(t *testing.T) {
@@ -649,7 +649,7 @@ func TestEditCodeMonitor(t *testing.T) {
 
 	// Update the code monitor.
 	// We update all fields, delete one action, and add a new action.
-	schema, err := graphqlbackend.NewSchema(db, nil, nil, nil, nil, r, nil, nil, nil, nil, nil)
+	schema, err := graphqlbackend.NewSchema(db, nil, nil, nil, nil, r, nil, nil, nil, nil, nil, nil)
 	require.NoError(t, err)
 	updateInput := map[string]interface{}{
 		"monitorID": string(relay.MarshalID(MonitorKind, 1)),
