+++ enterprise/cmd/frontend/internal/notebooks/resolvers/stars_resolvers_test.go
@@ -77,7 +77,7 @@ func TestCreateAndDeleteNotebookStars(t *testing.T) {
 	createdNotebooks := createNotebooks(t, db, []*notebooks.Notebook{notebookFixture(user1.ID, true), notebookFixture(user1.ID, false)})
 
 	database := database.NewDB(db)
-	schema, err := graphqlbackend.NewSchema(database, nil, nil, nil, nil, nil, nil, nil, nil, nil, NewResolver(database))
+	schema, err := graphqlbackend.NewSchema(database, nil, nil, nil, nil, nil, nil, nil, nil, nil, NewResolver(database), nil)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -154,7 +154,7 @@ func TestListNotebookStars(t *testing.T) {
 	}
 
 	database := database.NewDB(db)
-	schema, err := graphqlbackend.NewSchema(database, nil, nil, nil, nil, nil, nil, nil, nil, nil, NewResolver(database))
+	schema, err := graphqlbackend.NewSchema(database, nil, nil, nil, nil, nil, nil, nil, nil, nil, NewResolver(database), nil)
 	if err != nil {
 		t.Fatal(err)
 	}
