+++ enterprise/cmd/frontend/internal/orgrepos/resolvers/resolvers_test.go
@@ -47,7 +47,7 @@ func TestOrgRepositories(t *testing.T) {
 			Schema: func() *graphql.Schema {
 				t.Helper()
 
-				parsedSchema, parseSchemaErr := graphqlbackend.NewSchema(db, nil, nil, nil, nil, nil, nil, nil, nil, NewResolver(db), nil)
+				parsedSchema, parseSchemaErr := graphqlbackend.NewSchema(db, nil, nil, nil, nil, nil, nil, nil, nil, NewResolver(db), nil, nil)
 				if parseSchemaErr != nil {
 					t.Fatal(parseSchemaErr)
 				}
