+++ enterprise/internal/compute/command.go
