+++ enterprise/internal/compute/output_command_test.go
