+++ enterprise/internal/compute/replace_command.go
