+++ enterprise/internal/compute/replace_command_test.go
