+++ enterprise/internal/compute/result.go
