+++ enterprise/internal/compute/template.go
