+++ enterprise/internal/compute/template_test.go
